feat: add food selector to thieving page
Extract CompactFoodSelector from combat.dart into a shared widget so
players can view, switch, and eat food while thieving.
